new update

+++ app/src/main/java/com/android/documentationrecordviafingerprint/view/AppSettingsActivity.java
@@ -1,6 +1,7 @@
 package com.android.documentationrecordviafingerprint.view;
 
 import android.content.Context;
+import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.AdapterView;
@@ -11,9 +12,14 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import com.android.documentationrecordviafingerprint.R;
 import com.android.documentationrecordviafingerprint.controller.MyFirebaseDatabase;
+import com.android.documentationrecordviafingerprint.helper.IMyConstants;
+import com.android.documentationrecordviafingerprint.internetchecking.CheckInternetConnectivity;
+import com.android.documentationrecordviafingerprint.uihelper.CustomConfirmDialog;
+import com.google.android.material.snackbar.Snackbar;
 
-public class AppSettingsActivity extends AppCompatActivity {
+public class AppSettingsActivity extends AppCompatActivity implements IMyConstants {
     private Context context;
+    private CustomConfirmDialog customConfirmDialog;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -38,15 +44,51 @@ public class AppSettingsActivity extends AppCompatActivity {
 
                         break;
                     case 3:
-                        MyFirebaseDatabase.deleteAllUserData(AppSettingsActivity.this);
+                        customConfirmDialog = new CustomConfirmDialog(context, "WARNING!!!" +
+                                "\n\nAll your data will be deleted and can not be recovered!" +
+                                "\n\nAre you sure to delete all your data?");
+                        customConfirmDialog.setBtnText("Delete All Data")
+                                .dangerBtn()
+                                .setOkBtn(new View.OnClickListener() {
+                                    @Override
+                                    public void onClick(View v) {
+                                        if (CheckInternetConnectivity.isInternetConnected(context)) {
+                                            MyFirebaseDatabase.deleteUserAllData(AppSettingsActivity.this);
+                                        } else {
+                                            Snackbar.make(v, NO_INTERNET_CONNECTION, Snackbar.LENGTH_LONG).show();
+                                        }
+                                        customConfirmDialog.dismissDialog();
+                                    }
+                                });
                         break;
                     case 4:
-                        MyFirebaseDatabase.deleteUserAccount(AppSettingsActivity.this);
+                        customConfirmDialog = new CustomConfirmDialog(context, "WARNING!!!" +
+                                "\n\nAll your data will be deleted and can not be recovered!" +
+                                "\n\nAre you sure to delete your account?");
+                        customConfirmDialog.setBtnText("Delete Account")
+                                .dangerBtn()
+                                .setOkBtn(new View.OnClickListener() {
+                                    @Override
+                                    public void onClick(View v) {
+                                        if (CheckInternetConnectivity.isInternetConnected(context)) {
+                                            MyFirebaseDatabase.deleteUserAccount(AppSettingsActivity.this);
+                                        } else {
+                                            Snackbar.make(v, NO_INTERNET_CONNECTION, Snackbar.LENGTH_LONG).show();
+                                        }
+                                        customConfirmDialog.dismissDialog();
+                                    }
+                                });
                         break;
                     default:
                         break;
                 }
             }
         });
+    }
+
+    @Override
+    public void onBackPressed() {
+        startActivity(new Intent(context, DashboardActivity.class));
+        finish();
     }
 }
